feat: enhance home page and rebrand to 'The Git Platform for AI Agents'
- Add comprehensive landing page with hero, features, trending repos, quick start guide
- Replace 'GitHub for AI Agents' with 'The Git Platform for AI Agents' to avoid trademark issues
- Update all documentation, SDKs, specs, and source files with new branding
- Fix admin login 401 error (SHA256 hash instead of bcrypt)
- Fix audit log and reconciliation API response transformations
- Fix React navigate() warning in AdminLoginPage
- Add /admin proxy route to Vite config
- Add author credit: James Karanja Maina

+++ backend/src/services/reconciliation.rs
@@ -137,7 +137,7 @@ impl ReconciliationService {
             let drift = DriftType::StarCountMismatch {
                 repo_id: row.repo_id.clone(),
                 expected: row.actual_count,
-                actual: row.expected_count,
+                actual: i64::from(row.expected_count),
             };
 
             // Log the drift
@@ -506,7 +506,7 @@ impl ReconciliationJob {
 #[derive(Debug, sqlx::FromRow)]
 struct StarCountMismatchRow {
     repo_id: String,
-    expected_count: i64,
+    expected_count: i32,
     actual_count: i64,
 }
